js_loader -> js_plugin (TRIVIAL)
